updated certificate

+++ src/modules/certificates/certificate-main/components/create/CertificateMainAdd.js
@@ -106,6 +106,7 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
     console.log("files", files[0]);
 
     if (files.length > 0) {
+      // setValue("multipleAttachments", "")
       files.forEach((file) => {
         const filesUpdated = [
           file,
@@ -169,6 +170,7 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
                         option.value
                       );
                       setValue("intCategoryID", "");
+                      setValue("intCertificateID", "");
                       dispatch(getCertificateChildCategoryData(option.value));
                       dispatch(handleCertificateCategoryInput("certificateCategoryParent", { label: option.label, value: option.value }));
                       dispatch(handleCertificateCategoryInput("intParentsCategoryID", option.value));
@@ -331,7 +333,7 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
                   />
                 </div> */}
 
-                <div className="col-lg-3 col-md-4">
+                {/* <div className="col-lg-3 col-md-4">
                   <label className="form-label mt-2 formFont">Code</label>
                   <Form.Control
                     type="text"
@@ -349,12 +351,12 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
                       maxLength: 100,
                     })}
                   />
-                  {/* <div className="inputError margin-minus-8">
-                    {errors.strCustomeCode &&
-                      errors.strCustomeCode.type === "required" &&
-                      "Certificate Code can't be blank"}
-                  </div> */}
-                </div>
+                  //  <div className="inputError margin-minus-8">
+                  //   {errors.strCustomeCode &&
+                  //     errors.strCustomeCode.type === "required" &&
+                  //     "Certificate Code can't be blank"}
+                  // </div> 
+                </div> */}
 
                 <div className="col-lg-3 col-md-4">
                   <label className="form-label mt-2 formFont">
@@ -880,6 +882,9 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
                       name="multipleAttachments"
                       multiple={true}
                       onDone={getFiles.bind(this)}
+                    // ref={ref=> this.fileInput = ref} 
+                    // register={register}
+                    // setValue={setValue}
                     />
                   </div>
                 </div>
@@ -921,9 +926,14 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
                                   {/* <i className="fa fa-edit text-success mr-2"></i> */}
                                   <i
                                     className="fa fa-trash text-danger pointer"
-                                    onClick={() =>
-                                      deleteMultipleAttachmentData(index)
-                                    }
+                                    onClick={() => {
+                                      if (
+                                        window.confirm(
+                                          "Are you sure you wish to delete this item?"
+                                        )
+                                      )
+                                        deleteMultipleAttachmentData(index)
+                                    }}
                                   ></i>
                                 </td>
                               </tr>
@@ -936,7 +946,7 @@ const CertificateMainAdd = withRouter(({ history, props }) => {
               </div>
               <div className="form-group row">
                 <div className="col-sm-10">
-                  <a onClick={() => { history.push("/certificates-main/list") }}>
+                  <a href onClick={() => { history.push("/certificates-main/list") }}>
                     <button type="button" className="cancelButton btn mr-3"> Back </button>
                   </a>
